refactor: implement hook useFavorites
